model:group/friend

+++ model/table.sql
@@ -4,7 +4,7 @@ USE burgerChat;
 
 -- name 考虑去unique
 CREATE table User(
-    id int NOT NULL AUTO_INCREMENT,
+    id BIGINT NOT NULL AUTO_INCREMENT,
     name VARCHAR(50) NOT NULL,
     password VARCHAR(50) NOT NULL,
     state ENUM('online', 'offline') DEFAULT 'offline',
@@ -13,14 +13,14 @@ CREATE table User(
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 CREATE table Friend(
-    userid INT NOT NULL,
-    friendid INT NOT NULL,
+    userid BIGINT NOT NULL,
+    friendid BIGINT NOT NULL,
     PRIMARY KEY(userid, friendid)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 -- name 考虑去unique
 CREATE table AllGroup(
-    id INT AUTO_INCREMENT,
+    id BIGINT AUTO_INCREMENT,
     groupName VARCHAR(50) NOT NULL,
     groupDesc VARCHAR(200) DEFAULT '', 
     UNIQUE(groupName),
@@ -29,13 +29,13 @@ CREATE table AllGroup(
 
 -- 多对多，要产生一个中间表
 CREATE table GroupUser(
-    groupid INT NOT NULL,
-    userid INT NOT NULL,
+    groupid BIGINT NOT NULL,
+    userid BIGINT NOT NULL,
     groupRole ENUM('creator', 'manager', 'normal') DEFAULT 'normal', 
     PRIMARY KEY(groupid, userid)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 CREATE table OfflineMsg(
-    userid INT NOT NULL,
+    userid BIGINT NOT NULL,
     message VARCHAR(500) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
